+++ src/components/MediatedRequestsActivities/components/RequesterInformation/RequesterInformation.js
@@ -19,7 +19,7 @@ import {
   memoizeValidation,
 } from '../../../../utils';
 import {
-  REQUEST_FORM_FIELD_NAMES,
+  MEDIATED_REQUEST_FORM_FIELD_NAMES,
   RESOURCE_KEYS,
   ENTER_EVENT_KEY,
   BASE_SPINNER_PROPS,
@@ -134,7 +134,7 @@ class RequesterInformation extends Component {
       this.setState({ validatedBarcode: null });
     }
 
-    form.change(REQUEST_FORM_FIELD_NAMES.REQUESTER_BARCODE, barcode);
+    form.change(MEDIATED_REQUEST_FORM_FIELD_NAMES.REQUESTER_BARCODE, barcode);
   };
 
   handleBlur = (input) => () => {
@@ -205,8 +205,8 @@ class RequesterInformation extends Component {
                     <Field
                       data-testid="requesterBarcodeField"
                       key={key}
-                      name={REQUEST_FORM_FIELD_NAMES.REQUESTER_BARCODE}
-                      validate={this.validate(REQUEST_FORM_FIELD_NAMES.REQUESTER_BARCODE, key)}
+                      name={MEDIATED_REQUEST_FORM_FIELD_NAMES.REQUESTER_BARCODE}
+                      validate={this.validate(MEDIATED_REQUEST_FORM_FIELD_NAMES.REQUESTER_BARCODE, key)}
                       validateFields={[]}
                     >
                       {({ input, meta }) => {