+++ src/utils.test.js
@@ -31,12 +31,12 @@ import {
 } from './utils';
 import {
   FULFILMENT_TYPES,
-  REQUEST_TYPE_TRANSLATIONS,
-  REQUEST_TYPES,
+  MEDIATED_REQUEST_TYPE_TRANSLATION_KEYS,
+  MEDIATED_REQUEST_TYPES,
   DEFAULT_VIEW_VALUE,
   ID_TYPE_MAP,
-  REQUEST_TYPE_ERROR_TRANSLATIONS,
-  REQUEST_TYPE_ERRORS,
+  MEDIATED_REQUEST_TYPE_ERROR_TRANSLATIONS,
+  MEDIATED_REQUEST_TYPE_ERROR_LEVEL,
 } from './constants';
 
 describe('utils', () => {
@@ -525,12 +525,12 @@ describe('utils', () => {
   describe('getRequestTypesOptions', () => {
     it('should return array of options', () => {
       const requestTypes = {
-        [REQUEST_TYPES.HOLD]: [],
+        [MEDIATED_REQUEST_TYPES.HOLD]: [],
       };
       const expectedResult = [
         {
-          id: REQUEST_TYPE_TRANSLATIONS[REQUEST_TYPES.HOLD],
-          value: REQUEST_TYPES.HOLD,
+          id: MEDIATED_REQUEST_TYPE_TRANSLATION_KEYS[MEDIATED_REQUEST_TYPES.HOLD],
+          value: MEDIATED_REQUEST_TYPES.HOLD,
         }
       ];
 
@@ -638,11 +638,11 @@ describe('utils', () => {
 
   describe('getNoRequestTypeErrorMessageId', () => {
     it('should return error for title level request', () => {
-      expect(getNoRequestTypeErrorMessageId(true)).toBe(REQUEST_TYPE_ERROR_TRANSLATIONS[REQUEST_TYPE_ERRORS.TITLE_LEVEL_ERROR]);
+      expect(getNoRequestTypeErrorMessageId(true)).toBe(MEDIATED_REQUEST_TYPE_ERROR_TRANSLATIONS[MEDIATED_REQUEST_TYPE_ERROR_LEVEL.TITLE_LEVEL_ERROR]);
     });
 
     it('should return error for item level request', () => {
-      expect(getNoRequestTypeErrorMessageId(false)).toBe(REQUEST_TYPE_ERROR_TRANSLATIONS[REQUEST_TYPE_ERRORS.ITEM_LEVEL_ERROR]);
+      expect(getNoRequestTypeErrorMessageId(false)).toBe(MEDIATED_REQUEST_TYPE_ERROR_TRANSLATIONS[MEDIATED_REQUEST_TYPE_ERROR_LEVEL.ITEM_LEVEL_ERROR]);
     });
   });